zlideT2/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
boto3==1.34.92
botocore==1.34.92
cachetools==5.3.3
cement==3.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
cryptography==42.0.5
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.11
django-allauth==0.61.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-otp==1.5.0
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.2
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
inflection==0.5.1
jmespath==1.0.1
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
lxml==5.1.0
nltk==3.8.1
oauth2==1.9.0.post1
oauthlib==3.2.2
openai==1.16.2
packaging==24.0
pathspec==0.10.1
pillow==10.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-pptx==0.6.23
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2023.12.25
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
semantic-version==2.8.5
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
termcolor==2.4.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.1.9
whitenoise==6.6.0
XlsxWriter==3.2.0
